Add files via upload
This HTML file can be used in Module 4 section 2.

"To carry out an SQL injection attack on your database, you will establish a web interface for the data. We have provided a basic web interface for you, but you also have the option to create your own if you prefer. "